bux fix can click on other map buttons while transitionsing

+++ MonsterCare Mechanics Prototype/Assets/_Scripts/CombatPrototype/New Monster Battle Scripts/Dmg_BodyParts.cs
@@ -9,7 +9,7 @@ public class Dmg_BodyParts : MonoBehaviour
     GameObject Overlay;
 
     bool currentlyAttacking;
-    public int DmgModifier;
+    public float DmgModifier;
     float Counter;
     bool tapped;
